Consolidate chapter launching on the homepage
- Move link and file chapter flows into a shared homepage launcher
- Redirect the legacy reader route and update navigation links
- Add homepage source requests and reader consolidation coverage
