+++ src/components/Header.jsx
@@ -133,7 +133,7 @@ const Header = () => {
         <div className="flex items-center gap-3">
           <Show when="signed-out">
             <Link
-              href="/reader"
+              href="/#start-reader"
               className="bg-ink px-4 py-2 font-mono text-[11px] font-semibold uppercase tracking-[0.14em] text-white transition hover:bg-ink-soft"
             >
               Read free
